Migrate all 140 scalar fields to direct field access (u8)
Second swing of the u8 reimplementation: drop every byte_field! accessor and
convert all call sites (src + tests + bins) to direct struct-field access
(state.player_x, state.frame_counter, ...). Boundary casts to/from i32 locals,
ROM reads, and address math were inserted mechanically from rustc's type
spans (~1100 sites).

Behavioral regressions from u8 wrapping that the attract trace didn't cover,
caught by the test suite + the unused_comparisons lint and fixed:
  - loop counters inferred as u8 (let mut y = 9 assigned to a u8 field) made
    `y < 0` never true -> infinite loop; annotated i32 (find_*_overlap).
  - (room_metadef_hi + 3) > 255 always false once u8 wraps -> widened operand.
  - field << N shifted in u8 before widening, dropping the carry into a 16-bit
    address (scratch2 << 2); now every shift widens its operand to i32 first
    (byte-intended sites re-truncate identically via the assignment cast).

Verified byte-exact: 800-frame RAM trace identical to baseline, 34 test bins
green, unused_comparisons clean, 240f play smoke clean.

+++ src/engine.rs
@@ -294,8 +294,8 @@ impl Engine {
     }
 
     pub fn prg_map_shadow(&mut self) {
-        self.ppu_map_prg(32768, self.state.prg_bank_8000());
-        self.ppu_map_prg(40960, self.state.prg_bank_a000());
+        self.ppu_map_prg(32768, ((self.state.prg_bank_8000) as i32));
+        self.ppu_map_prg(40960, ((self.state.prg_bank_a000) as i32));
     }
 
     pub fn load_ines(&mut self, rom: &[u8], init_ram_pattern: bool) -> Result<(), String> {
